move ShuffleStrings to pkg/proxy
