+++ pkg/util/slice/BUILD
@@ -10,7 +10,6 @@ go_library(
     name = "go_default_library",
     srcs = ["slice.go"],
     importpath = "k8s.io/kubernetes/pkg/util/slice",
-    deps = ["//staging/src/k8s.io/apimachinery/pkg/util/rand:go_default_library"],
 )
 
 go_test(
